+++ module/apmot/tracer_test.go
@@ -240,8 +240,8 @@ func TestCustomTags(t *testing.T) {
 	payloads := recorder.Payloads()
 	require.Len(t, payloads.Transactions, 1)
 	require.Len(t, payloads.Spans, 1)
-	assert.Equal(t, model.StringMap{{Key: "foo", Value: "bar"}}, payloads.Transactions[0].Context.Tags)
-	assert.Equal(t, model.StringMap{{Key: "baz", Value: "qux"}}, payloads.Spans[0].Context.Tags)
+	assert.Equal(t, model.IfaceMap{{Key: "foo", Value: "bar"}}, payloads.Transactions[0].Context.Tags)
+	assert.Equal(t, model.IfaceMap{{Key: "baz", Value: "qux"}}, payloads.Spans[0].Context.Tags)
 }
 
 func TestStartSpanFromContextMixed(t *testing.T) {